+++ src/AttributeResolverFactory/ReflectionAttributeResolverFactory.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Hydrator\ResolverFactory;
+namespace Yiisoft\Hydrator\AttributeResolverFactory;
 
 use ReflectionClass;
 use ReflectionException;
+use Yiisoft\Hydrator\AttributeResolverFactoryInterface;
 use Yiisoft\Hydrator\DataAttributeInterface;
 use Yiisoft\Hydrator\Exception\NonInstantiableException;
 use Yiisoft\Hydrator\ObjectFactory\ReflectionObjectFactory;
