+++ src/Hydrator.php
@@ -4,18 +4,13 @@ declare(strict_types=1);
 
 namespace Yiisoft\Hydrator;
 
-use ReflectionAttribute;
 use ReflectionClass;
 use ReflectionProperty;
 use Yiisoft\Hydrator\Exception\NonInstantiableException;
 use Yiisoft\Hydrator\Internal\ConstructorArgumentsExtractor;
-use Yiisoft\Hydrator\DataAttributesHandler;
-use Yiisoft\Hydrator\Internal\ObjectPropertiesFilter;
-use Yiisoft\Hydrator\ParameterAttributesHandler;
-use Yiisoft\Hydrator\ObjectFactory\ObjectFactoryInterface;
+use Yiisoft\Hydrator\Internal\ReflectionFilter;
 use Yiisoft\Hydrator\ObjectFactory\ReflectionObjectFactory;
-use Yiisoft\Hydrator\ResolverFactory\AttributeResolverFactoryInterface;
-use Yiisoft\Hydrator\ResolverFactory\ReflectionAttributeResolverFactory;
+use Yiisoft\Hydrator\AttributeResolverFactory\ReflectionAttributeResolverFactory;
 use Yiisoft\Hydrator\TypeCaster\CompositeTypeCaster;
 use Yiisoft\Hydrator\TypeCaster\HydratorTypeCaster;
 use Yiisoft\Hydrator\TypeCaster\PhpNativeTypeCaster;
@@ -27,16 +22,18 @@ use Yiisoft\Hydrator\TypeCaster\PhpNativeTypeCaster;
  */
 final class Hydrator implements HydratorInterface
 {
-    private ConstructorArgumentsExtractor $constructorArgumentsExtractor;
-    private DataAttributesHandler $dataAttributesHandler;
-    private ParameterAttributesHandler $parameterAttributesHandler;
-    private ObjectPropertiesFilter $objectPropertiesFilter;
-
-    private ObjectFactoryInterface $objectFactory;
     /**
      * @var TypeCasterInterface Type caster used to cast raw values.
      */
     private TypeCasterInterface $typeCaster;
+
+    private ObjectFactoryInterface $objectFactory;
+
+    private DataAttributesHandler $dataAttributesHandler;
+
+    private ParameterAttributesHandler $parameterAttributesHandler;
+
+    private ConstructorArgumentsExtractor $constructorArgumentsExtractor;
 
     /**
      * @param TypeCasterInterface|null $typeCaster Type caster used to cast raw values.
@@ -46,22 +43,21 @@ final class Hydrator implements HydratorInterface
         ?AttributeResolverFactoryInterface $attributeResolverFactory = null,
         ?ObjectFactoryInterface $objectFactory = null,
     ) {
-        $this->objectFactory = $objectFactory ?? new ReflectionObjectFactory();
-        $attributeResolverFactory ??= new ReflectionAttributeResolverFactory();
-
         $this->typeCaster = $typeCaster ?? new CompositeTypeCaster(
             new PhpNativeTypeCaster(),
             new HydratorTypeCaster(),
         );
 
+        $attributeResolverFactory ??= new ReflectionAttributeResolverFactory();
         $this->dataAttributesHandler = new DataAttributesHandler($attributeResolverFactory);
         $this->parameterAttributesHandler = new ParameterAttributesHandler($attributeResolverFactory);
-        $this->objectPropertiesFilter = new ObjectPropertiesFilter();
+
+        $this->objectFactory = $objectFactory ?? new ReflectionObjectFactory();
+
         $this->constructorArgumentsExtractor = new ConstructorArgumentsExtractor(
             $this,
             $this->parameterAttributesHandler,
             $this->typeCaster,
-            $this->objectPropertiesFilter,
         );
     }
 
@@ -69,13 +65,14 @@ final class Hydrator implements HydratorInterface
     {
         $dataObject = new Data($data, $map, $strict);
         $reflectionClass = new ReflectionClass($object);
+
         $this->dataAttributesHandler->handle($reflectionClass, $dataObject);
 
-        $reflectionProperties = $this->objectPropertiesFilter->filterReflectionProperties(
-            $reflectionClass->getProperties(),
-            []
+        $this->hydrateInternal(
+            $object,
+            ReflectionFilter::filterProperties($reflectionClass),
+            $dataObject
         );
-        $this->hydrateInternal($object, $reflectionProperties, $dataObject);
     }
 
     public function create(string $class, array $data = [], array $map = [], bool $strict = false): object
@@ -86,6 +83,7 @@ final class Hydrator implements HydratorInterface
 
         $dataObject = new Data($data, $map, $strict);
         $reflectionClass = new ReflectionClass($class);
+
         $this->dataAttributesHandler->handle($reflectionClass, $dataObject);
 
         [$excludeProperties, $constructorArguments] = $this->constructorArgumentsExtractor->extract(
@@ -93,13 +91,13 @@ final class Hydrator implements HydratorInterface
             $dataObject,
         );
 
-        $reflectionProperties = $this->objectPropertiesFilter->filterReflectionProperties(
-            $reflectionClass->getProperties(),
-            $excludeProperties
-        );
-
         $object = $this->objectFactory->create($reflectionClass, $constructorArguments);
-        $this->hydrateInternal($object, $reflectionProperties, $dataObject);
+
+        $this->hydrateInternal(
+            $object,
+            ReflectionFilter::filterProperties($reflectionClass, $excludeProperties),
+            $dataObject
+        );
 
         return $object;
     }
@@ -107,6 +105,8 @@ final class Hydrator implements HydratorInterface
     /**
      * @param array<string, ReflectionProperty> $reflectionProperties
      * @psalm-param MapType $map
+     *
+     * @throws NonInstantiableException
      */
     private function hydrateInternal(
         object $object,
