+++ src/HydratorInterface.php
@@ -22,6 +22,8 @@ interface HydratorInterface
      * an object.
      * @param bool $strict Whether to hydrate properties from the map only.
      *
+     * @throws NonInstantiableException
+     *
      * @psalm-param MapType $map
      */
     public function hydrate(object $object, array $data = [], array $map = [], bool $strict = false): void;
