+++ src/ObjectFactory/ReflectionObjectFactory.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Yiisoft\Hydrator\ObjectFactory;
 
 use ReflectionClass;
-use ReflectionException;
 use Yiisoft\Hydrator\Exception\NonInstantiableException;
+use Yiisoft\Hydrator\ObjectFactoryInterface;
 
 use function count;
 
@@ -18,7 +18,6 @@ final class ReflectionObjectFactory implements ObjectFactoryInterface
      * @psalm-return T
      *
      * @throws NonInstantiableException
-     * @throws ReflectionException
      */
     public function create(ReflectionClass $reflectionClass, array $constructorArguments): object
     {
