+++ tests/ParameterAttributesHandlerTest.php
@@ -7,8 +7,8 @@ namespace Yiisoft\Hydrator\Tests;
 use PHPUnit\Framework\TestCase;
 use Yiisoft\Hydrator\Context;
 use Yiisoft\Hydrator\ParameterAttributesHandler;
-use Yiisoft\Hydrator\ResolverFactory\ContainerAttributeResolverFactory;
-use Yiisoft\Hydrator\ResolverFactory\ReflectionAttributeResolverFactory;
+use Yiisoft\Hydrator\AttributeResolverFactory\ContainerAttributeResolverFactory;
+use Yiisoft\Hydrator\AttributeResolverFactory\ReflectionAttributeResolverFactory;
 use Yiisoft\Hydrator\Tests\Support\Attribute\ContextViewer;
 use Yiisoft\Hydrator\Tests\Support\Attribute\ContextViewerResolver;
 use Yiisoft\Hydrator\Tests\Support\Attribute\CustomValue;
